+++ torchbearer/callbacks/__init__.py
@@ -19,6 +19,6 @@ from .live_loss_plot import LiveLossPlot
 from . import init
 from . import imaging
 from .pycm import PyCM
-from .mixup import Mixup
+from .mixup import Mixup, MixupAcc
 from .sample_pairing import SamplePairing
 from .label_smoothing import LabelSmoothingRegularisation